Rely on `ActiveSupport::Notifications` and `ActiveSupport::LogSubscriber` for logging and instrumentation

+++ app/models/solid_queue/execution/dispatching.rb
@@ -9,9 +9,11 @@ module SolidQueue
         def dispatch_jobs(job_ids)
           jobs = Job.where(id: job_ids)
 
-          Job.dispatch_all(jobs).map(&:id).tap do |dispatched_job_ids|
-            where(job_id: dispatched_job_ids).order(:job_id).delete_all
-            SolidQueue.logger.info("[SolidQueue] Dispatched #{dispatched_job_ids.size} jobs")
+          Job.dispatch_all(jobs).map(&:id).then do |dispatched_job_ids|
+            if dispatched_job_ids.none? then 0
+            else
+              where(job_id: dispatched_job_ids).order(:job_id).delete_all
+            end
           end
         end
       end
